Add async return type nullability fix and update tests
