+++ Nullness.Bang.Tests/CS8603CodeFixTests.cs
@@ -39,5 +39,41 @@ namespace Nullness.Bang.Tests
 
             await testhost.RunAsync();
         }
+
+        [TestMethod]
+        public async Task CS8603CodeFixTests_AsyncShould_UpdateReturnTypeToNullable()
+        {
+            var testhost = new BangCodeFixTestHost()
+            {
+                TestCode = """ 
+                #nullable enable
+                using System.Threading.Tasks;
+                
+                public class X {
+                    public async Task<string> demo()
+                    {
+                        return null;
+                    } 
+                }
+                """,
+                FixedCode = """ 
+                #nullable enable
+                using System.Threading.Tasks;
+                
+                public class X {
+                    public async Task<string?> demo()
+                    {
+                        return null;
+                    } 
+                }
+                """,
+            };
+
+            testhost.ExpectedDiagnostics.Add(
+                DiagnosticResult.CompilerWarning("CS8603").WithSpan(7, 16, 7, 20)
+            );
+
+            await testhost.RunAsync();
+        }
     }
 }
